Changes to underlying structure
-Rewrote Game, Board, and Player so that they're contained within
themselves. For example, the game class just controls when each player
takes their turn, all of the checking and the code for updating the
board has been refactored back into their respective classes. The Game
class is no longer doing something that the Player class should be doing
instead.
-Added a few helper functions and a few get/set functions as were
necessary in each of the classes
-Adjusted the main to take into account the information that is coming
out of the game object.
-The game is buggy as all get out, but the code compiles and it runs to
a certain degree.